+++ packages/zowe-explorer-api/src/vscode/ZoweVsCodeExtension.ts
@@ -12,6 +12,7 @@
 import * as semver from "semver";
 import * as vscode from "vscode";
 import { ZoweExplorerApi } from "../profiles";
+import { IZoweLogger, MessageSeverityEnum } from "../logger/IZoweLogger";
 
 /**
  * Collection of utility functions for writing Zowe Explorer VS Code extensions.
@@ -35,5 +36,20 @@ export class ZoweVsCodeExtension {
             return zoweExplorerApi.exports as ZoweExplorerApi.IApiRegisterClient;
         }
         return undefined;
+    }
+
+    /**
+     * Reveal an error in VSCode, and log the error to the Imperative log
+     *
+     */
+    public static showVsCodeMessage(message: string, severity: MessageSeverityEnum, logger: IZoweLogger): void {
+        logger.logImperativeMessage(message, severity);
+
+        const errorMessage = `${logger.getExtensionName()} error: ${message}`;
+        if (severity < 3) {
+            void vscode.window.showInformationMessage(errorMessage);
+        } else {
+            void vscode.window.showErrorMessage(errorMessage);
+        }
     }
 }