+++ packages/zowe-explorer/src/Profiles.ts
@@ -33,6 +33,7 @@ import {
     IValidationSetting,
     ValidProfileEnum,
     ProfilesCache,
+    IUrlValidator,
 } from "@zowe/zowe-explorer-api";
 import { errorHandling, FilterDescriptor, FilterItem, resolveQuickPickHelper, isTheia } from "./utils/ProfilesUtils";
 import { ZoweExplorerApiRegister } from "./ZoweExplorerApiRegister";
@@ -280,10 +281,21 @@ export class Profiles extends ProfilesCache {
         const createPick = new FilterDescriptor("\uFF0B " + createNewProfile);
         const items: vscode.QuickPickItem[] = profileNamesList.map((element) => new FilterItem(element));
         const quickpick = vscode.window.createQuickPick();
-        const placeholder = localize(
+        let placeholder = localize(
             "addSession.quickPickOption",
-            'Choose "Create new..." to define a new profile or select an existing profile to Add to the USS Explorer'
+            'Choose "Create new..." to define or select a profile to add to the USS Explorer'
         );
+        if (zoweFileProvider.getTreeType() === PersistenceSchemaEnum.Dataset) {
+            placeholder = localize(
+                "addSession.quickPickOption",
+                'Choose "Create new..." to define or select a profile to add to the DATA SETS Explorer'
+            );
+        } else if (zoweFileProvider.getTreeType() === PersistenceSchemaEnum.Job) {
+            placeholder = localize(
+                "addSession.quickPickOption",
+                'Choose "Create new..." to define or select a profile to add to the Job Views Explorer'
+            );
+        }
 
         if (isTheia()) {
             const options: vscode.QuickPickOptions = {
@@ -373,7 +385,7 @@ export class Profiles extends ProfilesCache {
         let updUser: string;
         let updPass: string;
         let updRU: boolean;
-        let updUrl: any;
+        let updUrl: IUrlValidator | undefined;
         let updPort: any;
 
         const schema: {} = this.getSchema(profileLoaded.type);
@@ -537,7 +549,7 @@ export class Profiles extends ProfilesCache {
         let newUser: string;
         let newPass: string;
         let newRU: boolean;
-        let newUrl: any;
+        let newUrl: IUrlValidator | undefined;
         let newPort: any;
 
         const newProfileName = profileName.trim();
@@ -752,23 +764,35 @@ export class Profiles extends ProfilesCache {
             return undefined;
         } else {
             try {
+                let cancel = false;
                 const updSession = await ZoweExplorerApiRegister.getMvsApi(loadProfile).getSession();
                 if (rePrompt) {
                     const saveButton = localize("promptCredentials.saveCredentials.button", "Save Credentials");
                     const doNotSaveButton = localize("promptCredentials.doNotSave.button", "Do Not Save");
                     const infoMsg = localize(
                         "promptCredentials.saveCredentials.infoMessage",
-                        "Save entered credentials for future use with profile: {0}? Saving credentials will update the local yaml file.",
+                        "Save entered credentials for future use with profile: {0}?\nSaving credentials will update the local yaml file.",
                         loadProfile.name
                     );
                     await vscode.window
-                        .showInformationMessage(infoMsg, ...[saveButton, doNotSaveButton])
-                        .then((selection) => {
+                        .showInformationMessage(infoMsg, { modal: true }, ...[saveButton, doNotSaveButton])
+                        .then(async (selection) => {
                             if (selection === saveButton) {
                                 rePrompt = false;
                             }
+                            if (!selection || selection === "Cancel") {
+                                cancel = true;
+                                vscode.window.showInformationMessage(
+                                    localize("promptCredentials.saveCredentials.cancelled", "Operation Cancelled")
+                                );
+                                await this.refresh(ZoweExplorerApiRegister.getInstance());
+                                return;
+                            }
                         });
                     await this.updateProfile(loadProfile, rePrompt);
+                }
+                if (cancel) {
+                    return undefined;
                 }
                 return [updSession.ISession.user, updSession.ISession.password, updSession.ISession.base64EncodedAuth];
             } catch (error) {
@@ -1381,59 +1405,54 @@ export class Profiles extends ProfilesCache {
 
     // ** Functions for handling Profile Information */
 
-    private async urlInfo(input?) {
+    private async urlInfo(input?): Promise<IUrlValidator | undefined> {
         let zosURL: string;
-
-        const urlInputBox = vscode.window.createInputBox();
         if (input) {
-            urlInputBox.value = input;
+            zosURL = input;
         }
-        urlInputBox.ignoreFocusOut = true;
-        urlInputBox.placeholder = localize("createNewConnection.option.prompt.url.placeholder", "https://url:port");
-        urlInputBox.prompt = localize(
-            "createNewConnection.option.prompt.url",
-            "Enter a z/OS URL in the format 'https://url:port'."
-        );
-
-        urlInputBox.show();
-        zosURL = await this.getUrl(urlInputBox);
-        urlInputBox.dispose();
-
-        if (!zosURL) {
-            return undefined;
-        }
-
-        return this.validateAndParseUrl(zosURL);
-    }
-
-    private async getUrl(urlInputBox): Promise<string | undefined> {
-        return new Promise<string | undefined>((resolve, reject) => {
-            urlInputBox.onDidHide(() => {
-                reject(undefined);
-                resolve(urlInputBox.value);
-            });
-            urlInputBox.onDidAccept(() => {
-                let host: string;
-                if (urlInputBox.value.includes(":")) {
-                    if (urlInputBox.value.includes("/")) {
-                        host = urlInputBox.value;
-                    } else {
-                        host = `https://${urlInputBox.value}`;
-                    }
-                } else {
-                    host = `https://${urlInputBox.value}`;
-                }
-
+        zosURL = await vscode.window.showInputBox({
+            prompt: localize(
+                "createNewConnection.option.prompt.url",
+                "Enter a z/OS URL in the format 'https://url:port'."
+            ),
+            value: zosURL,
+            ignoreFocusOut: true,
+            placeHolder: localize("createNewConnection.option.prompt.url.placeholder", "https://url:port"),
+            validateInput: (text: string): string | undefined => {
+                const host = this.getUrl(text);
                 if (this.validateAndParseUrl(host).valid) {
-                    resolve(host);
+                    return undefined;
                 } else {
-                    urlInputBox.validationMessage = localize(
+                    return localize(
                         "createNewConnection.invalidzosURL",
                         "Please enter a valid host URL in the format 'company.com'."
                     );
                 }
-            });
+            },
         });
+
+        let hostName: string;
+        if (!zosURL) {
+            return undefined;
+        } else {
+            hostName = this.getUrl(zosURL);
+        }
+
+        return this.validateAndParseUrl(hostName);
+    }
+
+    private getUrl(host: string): string {
+        let url: string;
+        if (host.includes(":")) {
+            if (host.includes("/")) {
+                url = host;
+            } else {
+                url = `https://${host}`;
+            }
+        } else {
+            url = `https://${host}`;
+        }
+        return url;
     }
 
     private async portInfo(input: string, schema: {}) {