+++ packages/zowe-explorer/src/ZoweExplorerExtender.ts
@@ -22,6 +22,7 @@ import {
     IZoweDatasetTreeNode,
     IZoweUSSTreeNode,
     IZoweJobTreeNode,
+    ProfilesCache,
 } from "@zowe/zowe-explorer-api";
 import { Profiles } from "./Profiles";
 import { getProfile, getLinkedProfile } from "./ProfileLink";
@@ -77,7 +78,16 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
         public jobsProvider?: IZoweTree<IZoweJobTreeNode>
     ) {}
 
-    public async initForZowe(type: string, meta: imperative.ICommandProfileTypeConfiguration[]) {
+    /**
+     *
+     * @implements ZoweExplorerApi.IApiExplorerExtender.initForZowe()
+     * @param {string} profileType
+     * @param {imperative.ICommandProfileTypeConfiguration[]} profileTypeConfigurations
+     */
+    public async initForZowe(
+        profileType: string,
+        profileTypeConfigurations: imperative.ICommandProfileTypeConfiguration[]
+    ) {
         // Ensure that when a user has not installed the profile type's CLI plugin
         // and/or created a profile that the profile directory in ~/.zowe/profiles
         // will be created with the appropriate meta data. If not called the user will
@@ -86,8 +96,8 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
             defaultHome: path.join(os.homedir(), ".zowe"),
             envVariablePrefix: "ZOWE",
         };
-        const configOptions = Array.from(meta);
-        const exists = fs.existsSync(path.posix.join(`${os.homedir}/.zowe/profiles/${type}`));
+        const configOptions = Array.from(profileTypeConfigurations);
+        const exists = fs.existsSync(path.posix.join(`${os.homedir()}/.zowe/profiles/${profileType}`));
         if (configOptions && !exists) {
             await imperative.CliProfileManager.initialize({
                 configuration: configOptions,
@@ -116,6 +126,7 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
      * profile types that can be employed in conjunction with the primary profile to provide
      * alternative support.
      *
+     * @deprecated
      * @param primaryNode represents the Tree item that is being used
      * @return The requested profile
      */
@@ -124,11 +135,22 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
     }
 
     /**
+     * Gives extenders access to the profiles loaded into memory by Zowe Explorer.
+     *
+     * @implements ZoweExplorerApi.IApiExplorerExtender.getProfilesCache()
+     * @returns {ProfilesCache}
+     */
+    public getProfilesCache(): ProfilesCache {
+        return Profiles.getInstance();
+    }
+
+    /**
      * After an extenders registered all its API extensions it
      * might want to request that profiles should get reloaded
      * to make them automatically appears in the Explorer drop-
      * down dialogs.
      *
+     * @implements ZoweExplorerApi.IApiExplorerExtender.reloadProfiles()
      * @param profileType optional profile type that the extender can specify
      */
     public async reloadProfiles(profileType?: string): Promise<void> {