+++ packages/zowe-explorer/src/command/TsoCommandHandler.ts
@@ -19,6 +19,7 @@ import { Profiles } from "../Profiles";
 import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
 import { errorHandling, FilterDescriptor, FilterItem, resolveQuickPickHelper } from "../utils/ProfilesUtils";
 import { ZoweCommandProvider } from "../abstract/ZoweCommandProvider";
+import { IStartTsoParms } from "@zowe/cli";
 
 // Set up localization
 nls.config({
@@ -50,9 +51,14 @@ export class TsoCommandHandler extends ZoweCommandProvider {
     private static readonly defaultDialogText: string =
         "\uFF0B " + localize("command.option.prompt.search", "Create a new TSO Command");
     private static instance: TsoCommandHandler;
+    public outputChannel: vscode.OutputChannel;
 
     constructor() {
         super();
+
+        this.outputChannel = vscode.window.createOutputChannel(
+            localize("issueTsoCommand.outputchannel.title", "Zowe TSO Command")
+        );
     }
 
     /**
@@ -128,9 +134,12 @@ export class TsoCommandHandler extends ZoweCommandProvider {
             if (Profiles.getInstance().validProfile !== ValidProfileEnum.INVALID) {
                 const commandApi = ZoweExplorerApiRegister.getInstance().getCommandApi(profile);
                 if (commandApi) {
-                    let acctNum: string;
+                    let tsoParams: IStartTsoParms;
                     if (profile.type === "zosmf") {
-                        acctNum = await this.getAccountNumber();
+                        tsoParams = await this.getTsoParams();
+                        if (!tsoParams) {
+                            return;
+                        }
                     }
                     let command1: string = command;
                     if (!command) {
@@ -138,7 +147,7 @@ export class TsoCommandHandler extends ZoweCommandProvider {
                             session && session.ISession ? session.ISession.hostname : "unknown"
                         );
                     }
-                    await this.issueCommand(command1, profile, acctNum);
+                    await this.issueCommand(command1, profile, tsoParams);
                 } else {
                     vscode.window.showErrorMessage(localize("issueTsoCommand.checkProfile", "Profile is invalid"));
                     return;
@@ -230,10 +239,11 @@ export class TsoCommandHandler extends ZoweCommandProvider {
     /**
      * Allow the user to submit an TSO command to the selected server. Response is written
      * to the output channel.
-     * @param session The Session object
      * @param command the command string
+     * @param profile profile to be used
+     * @param tsoParams parameters (from TSO profile, when used)
      */
-    private async issueCommand(command: string, profile: imperative.IProfileLoaded, acctNum?: string) {
+    private async issueCommand(command: string, profile: imperative.IProfileLoaded, tsoParams?: IStartTsoParms) {
         try {
             if (command) {
                 // If the user has started their command with a / then remove it
@@ -247,7 +257,20 @@ export class TsoCommandHandler extends ZoweCommandProvider {
                         title: localize("issueTsoCommand.command.submitted", "TSO command submitted."),
                     },
                     () => {
-                        return ZoweExplorerApiRegister.getCommandApi(profile).issueTsoCommand(command, acctNum);
+                        if (ZoweExplorerApiRegister.getCommandApi(profile).issueTsoCommandWithParms) {
+                            return ZoweExplorerApiRegister.getCommandApi(profile).issueTsoCommandWithParms(
+                                command,
+                                tsoParams
+                            );
+                        } else {
+                            let acctNum: string;
+                            if (!tsoParams) {
+                                acctNum = undefined;
+                            } else {
+                                acctNum = tsoParams.account;
+                            }
+                            return ZoweExplorerApiRegister.getCommandApi(profile).issueTsoCommand(command, acctNum);
+                        }
                     }
                 );
                 if (submitResponse.success) {
@@ -267,10 +290,10 @@ export class TsoCommandHandler extends ZoweCommandProvider {
         }
     }
 
-    private async getAccountNumber(): Promise<string> {
+    private async getTsoParams(): Promise<IStartTsoParms> {
         const tsoProfiles: imperative.IProfileLoaded[] = [];
         let tsoProfile: imperative.IProfileLoaded;
-        let acctNum: string;
+        const tsoParms: IStartTsoParms = {};
         const profileManager = Profiles.getInstance().getCliProfileManager("tso");
         if (profileManager) {
             try {
@@ -314,8 +337,16 @@ export class TsoCommandHandler extends ZoweCommandProvider {
             }
         }
         if (tsoProfile) {
-            acctNum = tsoProfile.profile.account;
+            tsoParms.account = tsoProfile.profile.account;
+            tsoParms.characterSet = tsoProfile.profile.characterSet;
+            tsoParms.codePage = tsoProfile.profile.codePage;
+            tsoParms.columns = tsoProfile.profile.columns;
+            tsoParms.logonProcedure = tsoProfile.profile.logonProcedure;
+            tsoParms.regionSize = tsoProfile.profile.regionSize;
+            tsoParms.rows = tsoProfile.profile.rows;
         } else {
+            // If there is no tso profile an account number is still required, so ask for one.
+            // All other properties of tsoParams will be undefined, so defaults will be utilized.
             const InputBoxOptions = {
                 placeHolder: localize("issueTsoCommand.command.account", "Account Number"),
                 prompt: localize(
@@ -323,14 +354,14 @@ export class TsoCommandHandler extends ZoweCommandProvider {
                     "Enter the account number for the TSO connection."
                 ),
                 ignoreFocusOut: true,
-                value: acctNum,
+                value: tsoParms.account,
             };
-            acctNum = await vscode.window.showInputBox(InputBoxOptions);
-            if (!acctNum) {
+            tsoParms.account = await vscode.window.showInputBox(InputBoxOptions);
+            if (!tsoParms.account) {
                 vscode.window.showInformationMessage(localize("issueTsoCommand.cancelled", "Operation Cancelled."));
                 return;
             }
         }
-        return acctNum;
+        return tsoParms;
     }
 }